1.0.86 (#91)
* 1.0.86

* wip

* wip

* Update HSQLUsageManager.kt

+++ core/src/main/kotlin/com/simiacryptus/skyenet/core/platform/file/UsageManager.kt
@@ -49,14 +49,14 @@ open class UsageManager(val root: File) : UsageInterface {
                                 Session(sessionId),
                                 User(email = user),
                                 modelEnum,
-                                com.simiacryptus.jopenai.ApiModel.Usage(prompt_tokens = value.toInt())
+                                com.simiacryptus.jopenai.ApiModel.Usage(prompt_tokens = value.toLong())
                             )
 
                             "output" -> incrementUsage(
                                 Session(sessionId),
                                 User(email = user),
                                 modelEnum,
-                                com.simiacryptus.jopenai.ApiModel.Usage(completion_tokens = value.toInt())
+                                com.simiacryptus.jopenai.ApiModel.Usage(completion_tokens = value.toLong())
                             )
 
                             "cost" -> incrementUsage(
